reorganize code and make methods private
